feat: normalizeString 함수 추가 및 eslint 설정 수정
normalizeString 함수를 helpers에 추가하여 문자열을 유니코드 NFC 정규화, 소문자 변환, 특수문자 제거 처리하도록 구현함.  
eslint의 optimize-regex 룰 문제로 normalize.js 파일을 eslint.config.js의 ignores에 추가하여 해당 파일에 대한 룰 적용을 회피함.  
기존 index.js에서 normalizeString 함수 내장 코드를 분리하여 모듈화함.
